Projets/02- Django/02-Pharmax/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
Pillow==10.1.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
